rewrite themer to use a react context
