+++ src/__tests__/ensureExports.spec.js
@@ -9,7 +9,9 @@ const expectedExports = [
   'GlobalTheme',
   'Normalize',
   'SetStyles',
-  'themer',
+  'Themer',
+  'ThemerContext',
+  'ThemerProvider',
   'withTheme',
   'themePropTypes',
   'responsive',
