+++ src/styles/themer/ThemerContext.d.ts
@@ -1,0 +1,9 @@
+import { Context } from 'react'
+import { Themer } from './Themer'
+
+declare interface ThemerContextInterface {
+  tick: number // for propogating theme updates through react
+  themer: Themer
+}
+
+export declare const ThemerContext: Context<ThemerContextInterface>
