+++ src/styles/themer/ThemerContext.js
@@ -1,0 +1,4 @@
+import { createContext } from 'react'
+import { Themer } from './Themer'
+
+export const ThemerContext = createContext({ themer: new Themer(), tick: 0 })
